# Conflicts:
#	.travis.yml
#	CHANGELOG.md
#	src/main/java/net/sf/jabref/JabRefMain.java
#	src/main/java/net/sf/jabref/gui/JabRefFrame.java
#	src/main/java/net/sf/jabref/gui/keyboard/KeyBindingPreferences.java
#	src/main/java/net/sf/jabref/gui/keyboard/KeyBindingsDialog.java
#	src/main/resources/l10n/JabRef_da.properties
#	src/main/resources/l10n/JabRef_de.properties
#	src/main/resources/l10n/JabRef_en.properties
#	src/main/resources/l10n/JabRef_es.properties
#	src/main/resources/l10n/JabRef_fa.properties
#	src/main/resources/l10n/JabRef_fr.properties
#	src/main/resources/l10n/JabRef_in.properties
#	src/main/resources/l10n/JabRef_it.properties
#	src/main/resources/l10n/JabRef_ja.properties
#	src/main/resources/l10n/JabRef_nl.properties
#	src/main/resources/l10n/JabRef_no.properties
#	src/main/resources/l10n/JabRef_pt_BR.properties
#	src/main/resources/l10n/JabRef_ru.properties
#	src/main/resources/l10n/JabRef_sv.properties
#	src/main/resources/l10n/JabRef_tr.properties
#	src/main/resources/l10n/JabRef_vi.properties
#	src/main/resources/l10n/JabRef_zh.properties

+++ src/main/java/net/sf/jabref/importer/ImportFormatReader.java
@@ -15,15 +15,9 @@
 */
 package net.sf.jabref.importer;
 
-import java.io.BufferedInputStream;
-import java.io.File;
-import java.io.FileInputStream;
 import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
-import java.io.Reader;
-import java.nio.charset.Charset;
-import java.nio.charset.StandardCharsets;
+import java.nio.file.Path;
+import java.nio.file.Paths;
 import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
@@ -61,7 +55,8 @@ public class ImportFormatReader {
     public static final String BIBTEX_FORMAT = "BibTeX";
 
     /**
-     * all import formats, in the default order of import formats
+     * All import formats.
+     * Sorted accordingly to {@link ImportFormat#compareTo}, which defaults to alphabetically by the name
      */
     private final SortedSet<ImportFormat> formats = new TreeSet<>();
 
@@ -114,14 +109,14 @@ public class ImportFormatReader {
      */
     private Optional<ImportFormat> getByCliId(String cliId) {
         for (ImportFormat format : formats) {
-            if (format.getCLIId().equals(cliId)) {
+            if (format.getId().equals(cliId)) {
                 return Optional.of(format);
             }
         }
         return Optional.empty();
     }
 
-    public List<BibEntry> importFromFile(String format, String filename, OutputPrinter status)
+    public ParserResult importFromFile(String format, Path file)
             throws IOException {
         Optional<ImportFormat> importer = getByCliId(format);
 
@@ -129,28 +124,7 @@ public class ImportFormatReader {
             throw new IllegalArgumentException("Unknown import format: " + format);
         }
 
-        return importFromFile(importer.get(), filename, status);
-    }
-
-    public List<BibEntry> importFromFile(ImportFormat importer, String filename, OutputPrinter status) throws IOException {
-        Objects.requireNonNull(importer);
-        Objects.requireNonNull(filename);
-        File file = new File(filename);
-
-        try (InputStream stream = new FileInputStream(file);
-                BufferedInputStream bis = new BufferedInputStream(stream)) {
-
-            bis.mark(Integer.MAX_VALUE);
-
-            if (!importer.isRecognizedFormat(bis)) {
-                throw new IOException("Wrong file format");
-            }
-        }
-
-        try (InputStream stream = new FileInputStream(file);
-                BufferedInputStream bis = new BufferedInputStream(stream)) {
-            return importer.importEntries(bis, status);
-        }
+        return importer.get().importDatabase(file, Globals.prefs.getDefaultEncoding());
     }
 
     /**
@@ -184,33 +158,11 @@ public class ImportFormatReader {
             sb.append(StringUtil.repeatSpaces(pad));
 
             sb.append(" : ");
-            sb.append(imFo.getCLIId());
+            sb.append(imFo.getId());
             sb.append('\n');
         }
 
         return sb.toString();
-    }
-
-
-
-    public static InputStreamReader getUTF8Reader(File f) throws IOException {
-        return getReader(f, StandardCharsets.UTF_8);
-    }
-
-    public static InputStreamReader getUTF16Reader(File f) throws IOException {
-        return getReader(f, StandardCharsets.UTF_16);
-    }
-
-    public static InputStreamReader getReader(File f, Charset charset)
-            throws IOException {
-        return new InputStreamReader(new FileInputStream(f), charset);
-    }
-
-    public static Reader getReaderDefaultEncoding(InputStream in) {
-        InputStreamReader reader;
-        reader = new InputStreamReader(in, Globals.prefs.getDefaultEncoding());
-
-        return reader;
     }
 
     public static class UnknownFormatImport {
@@ -225,6 +177,9 @@ public class ImportFormatReader {
         }
     }
 
+    public UnknownFormatImport importUnknownFormat(String filename) {
+        return importUnknownFormat(Paths.get(filename));
+    }
 
     /**
      * Tries to import a file by iterating through the available import filters,
@@ -234,25 +189,20 @@ public class ImportFormatReader {
      *
      * @throws IOException
      */
-    public UnknownFormatImport importUnknownFormat(String filename) {
-        Objects.requireNonNull(filename);
+    public UnknownFormatImport importUnknownFormat(Path file) {
+        Objects.requireNonNull(file);
 
         // First, see if it is a BibTeX file:
         try {
-            ParserResult pr = OpenDatabaseAction.loadDatabase(new File(filename),
+            ParserResult pr = OpenDatabaseAction.loadDatabase(file.toFile(),
                     Globals.prefs.getDefaultEncoding());
             if (pr.getDatabase().hasEntries() || !pr.getDatabase().hasNoStrings()) {
-                pr.setFile(new File(filename));
+                pr.setFile(file.toFile());
                 return new UnknownFormatImport(ImportFormatReader.BIBTEX_FORMAT, pr);
             }
         } catch (IOException ignore) {
             // Ignored
         }
-
-        // we don't use a provided OutputPrinter (such as the JabRef frame),
-        // as we don't want to see any outputs from failed importers:
-        // we expect failures and do not want to report them to the user
-        OutputPrinterToNull nullOutput = new OutputPrinterToNull();
 
         // stores ref to best result, gets updated at the next loop
         List<BibEntry> bestResult = null;
@@ -261,14 +211,16 @@ public class ImportFormatReader {
 
         // Cycle through all importers:
         for (ImportFormat imFo : getImportFormats()) {
-
             try {
+                if(!imFo.isRecognizedFormat(file, Globals.prefs.getDefaultEncoding())) {
+                    continue;
+                }
 
-                List<BibEntry> entries = importFromFile(imFo, filename, nullOutput);
+                ParserResult parserResult = imFo.importDatabase(file, Globals.prefs.getDefaultEncoding());
+                List<BibEntry> entries = parserResult.getDatabase().getEntries();
 
-                int entryCount;
                 BibDatabases.purgeEmptyEntries(entries);
-                entryCount = entries.size();
+                int entryCount = entries.size();
 
                 if (entryCount > bestResultCount) {
                     bestResult = entries;
@@ -276,7 +228,7 @@ public class ImportFormatReader {
                     bestFormatName = imFo.getFormatName();
                 }
             } catch (IOException ex) {
-                // The import didn't succeed. Go on.
+                // The import did not succeed. Go on.
             }
         }
